Revert "Issue 150 histogram sample chart"
