+++ front-end/package.json
@@ -14,13 +14,11 @@
     "dotenv": "^8.2.0",
     "eslint": "^7.23.0",
     "http-proxy-middleware": "^2.0.0",
-    "plotly.js": "^2.2.0",
     "react": "^17.0.1",
     "react-copy-to-clipboard": "^5.0.3",
     "react-dom": "^17.0.1",
     "react-icons": "^4.2.0",
     "react-modal": "^3.12.1",
-    "react-plotly.js": "^2.5.1",
     "react-router-dom": "^5.2.0",
     "react-scripts": "4.0.3",
     "react-select": "^4.3.0",
