+++ front-end/src/css/pages/SampleDataUse.css
@@ -1,0 +1,51 @@
+.full-img,
+.full-text {
+  display: block;
+  margin-top: 0;
+  margin-left: auto;
+  margin-right: auto;
+  width: 90%;
+}
+
+.header-text {
+  text-align: center;
+  color: blue;
+  -webkit-text-stroke: 1px black;
+}
+
+.block {
+  border: 1px solid black;
+  width: 90%;
+  margin-left: auto;
+  margin-right: auto;
+}
+
+@media screen and (max-width: 2400px) {
+  .full-text {
+    font-size: 16px;
+  }
+  .header-text {
+    font-size: 35.5px;
+  }
+}
+
+@media screen and (max-width: 992px) {
+  .full-text {
+    font-size: 14px;
+  }
+  .header-text {
+    font-size: 35px;
+  }
+}
+
+@media screen and (max-width: 650px) {
+  .full-text {
+    font-size: 12px;
+  }
+  .full-img {
+    width: 95%;
+  }
+  .header-text {
+    font-size: 22px;
+  }
+}
